Required fields should raise an error if missing
Required fields should raise an error if missing, rather than silently
initialising to a meaningless value. h/t @cmshobe .

I note there’s a second issue here that `topographic__gradient` doesn’t
appear to be an input field. It’s created if it’s missing, and even if
it isn’t missing, it automatically gets immediately overwritten during
instantiation with a new set of values. That sounds like an output
field to me.